Refactor: move helpers of use term to UseTermHelper
