+++ ideax/ideax/tests/views/test_misc.py
@@ -1,6 +1,6 @@
 import json
 
-from datetime import date, datetime, timedelta
+from datetime import datetime, timedelta
 from itertools import product
 
 from django.db.models import QuerySet
@@ -9,8 +9,7 @@ import pytz
 
 from model_mommy import mommy
 
-from ...views import (get_authors, get_category_list, get_featured_challenges, get_phases, get_term_of_user,
-                      get_use_term_list)
+from ...views import get_authors, get_category_list, get_featured_challenges, get_phases, get_term_of_user
 
 
 class TestNonMiscView:
@@ -40,14 +39,6 @@ class TestNonMiscView:
         response = get_term_of_user(request)
         assert response.status_code == 200
         assert json.loads(response.content) == {'term': 'EULA Test'}
-
-    def test_get_use_term_list(self, ideax_views, db, mocker):
-        datelib = mocker.patch('ideax.ideax.views.use_term.date')
-        datelib.today.return_value = date(2010, 1, 1)
-        response = get_use_term_list()
-        assert len(response['use_term_list']) == 1
-        assert response['use_term_list'][0].term == 'A generic Term of Use.'
-        assert response['today'] == date(2010, 1, 1)
 
     def test_get_featured_challenges_empty(self, db):
         response = get_featured_challenges()
